requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
werkzeug==3.0.1
openpyxl==3.1.2
